Create Dart_Filter.ipynb
Got information from Zachary Ross on how to filter. Thanks man!

"I used the ObsPy interpolate method to resample all data at 100 Hz, only for data that were > 100 Hz, which is generally very few. The de-trending is just fitting a linear trend to the data and subtracting it from the time series.

The Butterworth filter is a bandpass filter with the corners at 1 Hz and 20 Hz. 4 corners were used. (ObsPy signal.filter.bandpass)"